Registering users - M5
Added a new screen to accept information to register user. The full name entry can be removed, I just thought I remember something about a User having a name as well as username.
Added a constants class in case we need it. I used it for logging.
The UserHandler takes care of maintaining the list of users (stored in a text file) and checking login information. I suppose it could be done better and not write to the file as much, but that's something for later.